requirements.txt
# Core dependencies
openai>=2.1.0
snowflake-connector-python>=3.12.0
pandas>=2.2.0
numpy>=1.26.0
lxml>=5.3.0
pydantic>=2.9.0
tenacity>=9.0.0
tiktoken>=0.7.0
python-dotenv>=1.0.0
colorlog>=6.8.0
streamlit>=1.39.0

# Testing (optional)
pytest>=8.3.0
pytest-cov>=5.0.0
openpyxl>=3.1.0
xlrd>=2.0.0
streamlit>=1.28.0
